Draw a different conclusion
Shifting the emphasis to error reporting was a mistake.  Failure tolerance
isn't inherently bad, it's only bad when it is necessarily outside of the
proscribed space defined by specification.  Indeed, some specifications
(notably HTML) value tolerance of errors and provide specific guidance about
that tolerance.  That type of design ensures that fault tolerance is clearly
specified.

Where fault tolerance comes in is where it forms part of a feedback loop.  As
maintainers of protocols (as distinct from mere authors who discarge their
responsibility when the specification is "published"), the ability to continue
maintenace depends greatly on the amount and quality of feedback from
implementations and deployments.  Reporting faults is merely a part of that.

Closes #3, #2.