Add section on flexibility in design
There was something of a side note in the previous draft that mentioned
HTML.  This was easy to overlook and the discussion about routing
protocols made it clear that more substance was needed.

Job Snijders, who knows more about BGP than I ever hope to, was good
enough to point to RFC 7606, which is a great example of the BGP
community doing positive protocol maintenance.  This example is now the
backbone of a proposed new section on making flexible protocols.  This
section closes with a comparison between doing it right (as RFC 7606
does) and the wild west of the robustness principle.