Proposal to rephrase abstract and intro
Probably based on the history of this document, I think that the abstract and intro (at least) is too much focused on the bad sides of the robustness principle than the actual good recommends that this document which therefore might get easily lost for the reader. This PR is an attempt to refocus the scope of the abstract and intro to what the document actually wants to say or I believe it wants to say. Please take this as proposal for consideration and of course feel free to edit as you want!

Btw. I think the document is generally a bit the wrong way around. What I learnt about paper writing is that the important message should come first and the explanation follows - this is true on a per section, as well per paragraph, and per sentence level, e.g. in a section the first paragraph has the most important message and in each paragraph the first sentence contains the actual point and all other sentences are only for "remarks", i.e. ideally you can quick read a section by only reading the first sentence of each paragraph. Doing it the other ways around distracts the reader form the points you actually want to make, which I think is part of the problem here.