Tolerate the unexpected
The aspect of robustness we're attacking here can henceforth be referred
to as "tolerating the unexpected".  I've not used this everywhere as we
do often refer to the robustness principle as a unified concept.  Still,
where we refer to the narrow concept, this fits remarkably well.

I'm not 100% on some of the changes, but I think that this coming
together nicely.